refactor!: restructure & update resources/models
This commit restructures the app in anticipation of bringing in an admin
module.

This commit also updates the resource definitions and usage to reflect
that the calendar app user should have some data that is only scoped to
their user account. (see Calendar.tsx use of URLs that depend upon the
username.)