Feat: Replace Category drawer with a menu in place of the list of equipment.
Replaces the category drawer with a menu that replaces the list
of equipment. Navigation in the 2 modes is now the similar. The
nested list navigation mode remains the same, and now the category
mode presents the user with a list of nested categories/subcategories
and clicking a category or subcategory brings the user to a separate
page containing the gear in that category