feat: adds Snackbar component to Admin
Snackbar introduces a user feedback component to display action
success/failure.

The component should be sharable between different modules.
This commit adds Snackbar to Admin.  This will be added to Calendar.

Module state needs a queue of states for the snackbar.
The state handler (reducer) should enqueue snackbar states.
The snackbar dequeues when it is closed.

Snackbars take a type (currently just "success" or "failure") which
handles appearances, a message to display, and an autoHideDuration which
can be a number in milliseconds to automatically hide the snackbar or
`null` if you want the snackbar to display until user interaction.