feat: adds error page
This is a final state page: the app is broken if this page is shown.
It provides a link to tech support, friendly messages, and can display
an error passed to it by the reducer.

TODO, eventually we can add error boundaries and use this as the
fallback page.
https://reactjs.org/docs/error-boundaries.html

This commit adds the error page to the admin app.
TODO - add to Calendar module.