fix: (experimental) date util: inputs: no TZ info
Unsure if this is exactly where to apply, but the issue discovered
was that the semester dates (and possibly others) coming back from
the MySQL database had their TZ info (".000Z") in the string, causing
the date input to display locale time and send off UTC to the database.

Other possible solution is to always strip this ".000Z" when storing
date strings, e.g. in the resource constructors.