fix: closes detail on success
The Formik form was not getting the updated ID passed in. Rather than
address that particular issue, this commit avoids it completely by just
closing the details form, which doesn't seem to be bad, but certainly
open for discussion.

the only changed line was the new line 57.
otherwise formatting changed.