feat: adds basic socket.io broadcasting
Adds basic way to notify all other users that changes have been made.

Currently just causes a "please refresh" snackbar to appear.
This is temporary.

TODO:
On receiving a broadcast, the app should have the means to silently update
and/or give the user a button to update when convenient.