fix: do not conditionally "unrender" calendar
Fixes issue where title bar date and calendar day get out of sync.
This was happening when all locations were deselected.
The logic had the calendar replaced with a "select location" message.
When the calendar returned, the initial date was back to current date,
and not the last selected date.